# [internal] Make GIL usage explicit for `session_poll_workunits`